Create blog.html.twig
Hi. With this version of the blog.html.twig file a list of items is added to the main page, as in the original HTML5Up template. It would be better to use a item.html.twig instead of default.html.twig for the items. The blog.html.twig file call to partials/blog-list-item.html.twig to show the latest posted items.